DO NOT SUBMIT: Fig merge commit (go/fig-merge-review)
This is a Fig merge commit, meaning it has multiple diffbase parents.
Critique does not, however, know how to deal with multiple parents, so
we create this CL as a placeholder - it's not meant to be submitted,
and typically disappears altogether when the author submits one of the
parents and syncs, so there's no need to review it.

The parents are:
- cl/417873258: [mlir] Add unit test for disabling canonicalizer patterns (NFC)
- cl/417752624: [mlir] Separate Listener from PatternRewriter

The original description of this merge, as entered by the author, was:
Merge Listener and Canonicalizer Labels